Crud users

+++ controllers/userController.js
@@ -5,7 +5,9 @@ const prisma = new PrismaClient();
 export const getAllUsers = async (_, res) => {
   try {
     const users = await prisma.user.findMany();
+
     if (users.length === 0) return res.status(404).json("Não há usuários cadastrados!");
+
     return res.status(200).json(users);
   } catch (error) {
     return res.status(500).json(error.message);
@@ -15,9 +17,11 @@ export const getAllUsers = async (_, res) => {
 export const getByUserId = async (req, res) => {
   try {
     const user = await prisma.user.findUnique({
-      where: { id: req.params.id },
+      where: { Id: req.params.id },
     });
+
     if (!user) return res.status(404).json("Usuário não encontrado!");
+
     return res.status(200).json(user);
   } catch (error) {
     return res.status(500).json(error.message);
@@ -25,10 +29,43 @@ export const getByUserId = async (req, res) => {
 };
 
 export const createUser = async (req, res) => {
-  const { nome, email, cpf } = req.body;
+  const {
+    Nome,
+    Sobrenome, 
+    CPF,
+    Telefone,
+    Genero,
+    Email,
+    Senha,
+    DataNasc, 
+    CEP, 
+    Logradouro, 
+    Bairro, 
+    Numero, 
+    Complemento,
+    Estado, 
+    Cidade 
+  } = req.body;
+
   try {
     const newUser = await prisma.user.create({
-      data: { nome, email, cpf },
+      data: {
+        Nome,
+        Sobrenome, 
+        CPF,
+        Telefone,
+        Genero,
+        Email,
+        Senha,
+        DataNasc, 
+        CEP, 
+        Logradouro, 
+        Bairro, 
+        Numero, 
+        Complemento,
+        Estado, 
+        Cidade 
+      },
     });
     return res.status(201).json("Usuário criado com sucesso!");
   } catch (error) {
@@ -37,11 +74,52 @@ export const createUser = async (req, res) => {
 };
 
 export const updateUser = async (req, res) => {
-  const { nome, email, cpf } = req.body;
+  const {
+    Nome,
+    Sobrenome, 
+    CPF,
+    Telefone,
+    Genero,
+    Email,
+    Senha,
+    DataNasc, 
+    CEP, 
+    Logradouro, 
+    Bairro, 
+    Numero, 
+    Complemento,
+    Estado, 
+    Cidade 
+  } = req.body;
+
   try {
+    const user = await prisma.user.findUnique({
+      where: { Id: req.params.id },  
+    });
+
+    if (!user) {
+      return res.status(404).json("Usuário não encontrado!");
+    }
+
     await prisma.user.update({
-      where: { id: req.params.id },
-      data: { nome, email, cpf },
+      where: { Id: req.params.id },
+      data: {
+        Nome,
+        Sobrenome, 
+        CPF,
+        Telefone,
+        Genero,
+        Email,
+        Senha,
+        DataNasc, 
+        CEP, 
+        Logradouro, 
+        Bairro, 
+        Numero, 
+        Complemento,
+        Estado, 
+        Cidade 
+      },
     });
     return res.status(200).json("Usuário atualizado com sucesso!");
   } catch (error) {
@@ -51,11 +129,21 @@ export const updateUser = async (req, res) => {
 
 export const deleteUser = async (req, res) => {
   try {
-    await prisma.user.delete({
-      where: { id: req.params.id },
+    const user = await prisma.user.findUnique({
+      where: { Id: req.params.id },  
     });
+
+    if (!user) {
+      return res.status(404).json("Usuário não encontrado!");
+    }
+
+    await prisma.user.delete({
+      where: { Id: req.params.id },
+    });
+
     return res.status(200).json("Usuário deletado com sucesso!");
   } catch (error) {
     return res.status(500).json(error.message);
   }
 };
+
